flatpak: allow access to $HOME
The whole purpose of Tuhi is to save the files after downloading and the
flatpak sandbox quietly dropped any file we saved in the target directory.

We could use xdg-download or something as allowed directory but let's open up
home itself since we don't want to limit where the files can be saved.

Fixes #230

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>